clean up some easy parts of product repository
