+++ src/Adapter/Product/Repository/ProductRepository.php
@@ -112,62 +112,6 @@ class ProductRepository extends AbstractObjectModelRepository
     }
 
     /**
-     * Gets position product position in category
-     *
-     * @param ProductId $productId
-     * @param CategoryId $categoryId
-     *
-     * @return int|null
-     */
-    public function getPositionInCategory(ProductId $productId, CategoryId $categoryId): ?int
-    {
-        $qb = $this->connection->createQueryBuilder();
-        $qb->select('position')
-            ->from($this->dbPrefix . 'category_product')
-            ->where('id_product = :productId')
-            ->andWhere('id_category = :categoryId')
-            ->setParameter('productId', $productId->getValue())
-            ->setParameter('categoryId', $categoryId->getValue())
-        ;
-
-        $position = $qb->execute()->fetchOne();
-
-        if (!$position) {
-            return null;
-        }
-
-        return (int) $position;
-    }
-
-    /**
-     * Gets product price by provided shop
-     *
-     * @param ProductId $productId
-     * @param ShopId $shopId
-     *
-     * @return DecimalNumber|null
-     */
-    public function getPriceByShop(ProductId $productId, ShopId $shopId): ?DecimalNumber
-    {
-        $qb = $this->connection->createQueryBuilder();
-        $qb->select('price')
-            ->from($this->dbPrefix . 'product_shop')
-            ->where('id_product = :productId')
-            ->andWhere('id_shop = :shopId')
-            ->setParameter('productId', $productId->getValue())
-            ->setParameter('shopId', $shopId->getValue())
-        ;
-
-        $result = $qb->execute()->fetch();
-
-        if (!$result) {
-            return null;
-        }
-
-        return new DecimalNumber($result['price']);
-    }
-
-    /**
      * @param ProductId $productId
      *
      * @throws ProductNotFoundException
@@ -175,67 +119,6 @@ class ProductRepository extends AbstractObjectModelRepository
     public function assertProductExists(ProductId $productId): void
     {
         $this->assertObjectModelExists($productId->getValue(), 'product', ProductNotFoundException::class);
-    }
-
-    /**
-     * @param ProductId[] $productIds
-     *
-     * @throws ProductNotFoundException
-     */
-    public function assertAllProductsExists(array $productIds): void
-    {
-        //@todo: no shop association. Should it be checked here?
-        $ids = array_map(function (ProductId $productId): int {
-            return $productId->getValue();
-        }, $productIds);
-        $ids = array_unique($ids);
-
-        $qb = $this->connection->createQueryBuilder();
-        $qb->select('COUNT(id_product) as product_count')
-            ->from($this->dbPrefix . 'product')
-            ->where('id_product IN (:productIds)')
-            ->setParameter('productIds', $ids, Connection::PARAM_INT_ARRAY)
-        ;
-
-        $results = $qb->execute()->fetch();
-
-        if (!$results || (int) $results['product_count'] !== count($ids)) {
-            throw new ProductNotFoundException(
-                sprintf(
-                    'Some of these products do not exist: %s',
-                    implode(',', $ids)
-                )
-            );
-        }
-    }
-
-    /**
-     * @param ProductId $productId
-     * @param LanguageId $languageId
-     *
-     * @return array<array<string, string>>
-     *                                      e.g [
-     *                                      ['id_product' => '1', 'name' => 'Product name', 'reference' => 'demo15'],
-     *                                      ['id_product' => '2', 'name' => 'Product name2', 'reference' => 'demo16'],
-     *                                      ]
-     *
-     * @throws CoreException
-     */
-    public function getRelatedProducts(ProductId $productId, LanguageId $languageId): array
-    {
-        $this->assertProductExists($productId);
-        $productIdValue = $productId->getValue();
-
-        try {
-            $accessories = Product::getAccessoriesLight($languageId->getValue(), $productIdValue);
-        } catch (PrestaShopException $e) {
-            throw new CoreException(sprintf(
-                'Error occurred when fetching related products for product #%d',
-                $productIdValue
-            ));
-        }
-
-        return $accessories;
     }
 
     /**
@@ -255,39 +138,6 @@ class ProductRepository extends AbstractObjectModelRepository
         );
 
         return $this->loadProduct($product);
-    }
-
-    /**
-     * @param ProductId $productId
-     *
-     * @return ProductType
-     *
-     * @throws ProductNotFoundException
-     */
-    public function getProductType(ProductId $productId): ProductType
-    {
-        $result = $this->connection->createQueryBuilder()
-            ->select('p.product_type')
-            ->from($this->dbPrefix . 'product', 'p')
-            ->where('p.id_product = :productId')
-            ->setParameter('productId', $productId->getValue())
-            ->execute()
-            ->fetchAssociative()
-        ;
-
-        if (empty($result)) {
-            throw new ProductNotFoundException(sprintf(
-                'Cannot find product type for product %d because it does not exist',
-                $productId->getValue()
-            ));
-        }
-
-        if (!empty($result['product_type'])) {
-            return new ProductType($result['product_type']);
-        }
-
-        // Older products that were created before product page v2, might have no type, so we determine it dynamically
-        return new ProductType($this->get($productId)->getDynamicProductType());
     }
 
     /**
@@ -315,159 +165,6 @@ class ProductRepository extends AbstractObjectModelRepository
     public function delete(ProductId $productId): void
     {
         $this->deleteObjectModel($this->get($productId), CannotDeleteProductException::class);
-    }
-
-    /**
-     * @param string $searchPhrase
-     * @param LanguageId $languageId
-     * @param ShopId $shopId
-     * @param int|null $limit
-     *
-     * @return array<int, array<string, int|string>>
-     */
-    public function searchProducts(string $searchPhrase, LanguageId $languageId, ShopId $shopId, ?int $limit = null): array
-    {
-        $qb = $this->getSearchQueryBuilder(
-            $searchPhrase,
-            $languageId,
-            $shopId,
-            [],
-            $limit);
-        $qb
-            ->addSelect('p.id_product, pl.name, p.reference, i.id_image')
-            ->addGroupBy('p.id_product')
-            ->addOrderBy('pl.name', 'ASC')
-            ->addOrderBy('p.id_product', 'ASC')
-        ;
-
-        return $qb->execute()->fetchAllAssociative();
-    }
-
-    /**
-     * @param string $searchPhrase
-     * @param LanguageId $languageId
-     * @param ShopId $shopId
-     * @param array $filters
-     * @param int|null $limit
-     *
-     * @return array<int, array<string, int|string>>
-     *
-     * @throws Exception
-     * @throws ExceptionAlias
-     */
-    public function searchCombinations(
-        string $searchPhrase,
-        LanguageId $languageId,
-        ShopId $shopId,
-        array $filters = [],
-        ?int $limit = null
-    ): array {
-        $qb = $this->getSearchQueryBuilder(
-            $searchPhrase,
-            $languageId,
-            $shopId,
-            $filters,
-            $limit
-        );
-        $qb
-            ->addSelect('p.id_product, pa.id_product_attribute, pl.name, i.id_image')
-            ->addSelect('p.reference as product_reference')
-            ->addSelect('pa.reference as combination_reference')
-            ->addSelect('ai.id_image as combination_image_id')
-            ->leftJoin('p', $this->dbPrefix . 'product_attribute_image', 'ai', 'ai.id_product_attribute = pa.id_product_attribute')
-            ->addGroupBy('p.id_product, pa.id_product_attribute')
-            ->addOrderBy('pl.name', 'ASC')
-            ->addOrderBy('p.id_product', 'ASC')
-            ->addOrderBy('pa.id_product_attribute', 'ASC')
-        ;
-
-        return $qb->execute()->fetchAllAssociative();
-    }
-
-    public function getProductTaxRulesGroupId(ProductId $productId): TaxRulesGroupId
-    {
-        $result = $this->connection->createQueryBuilder()
-            ->addSelect('p.id_tax_rules_group')
-            ->from($this->dbPrefix . 'product', 'p')
-            ->where('id_product = :productId')
-            ->setParameter('productId', $productId->getValue())
-            ->execute()
-            ->fetchOne()
-        ;
-
-        return new TaxRulesGroupId((int) $result);
-    }
-
-    /**
-     * @param string $searchPhrase
-     * @param LanguageId $languageId
-     * @param ShopId $shopId
-     * @param array $filters
-     * @param int|null $limit
-     *
-     * @return QueryBuilder
-     */
-    protected function getSearchQueryBuilder(
-        string $searchPhrase,
-        LanguageId $languageId,
-        ShopId $shopId,
-        array $filters = [],
-        ?int $limit = null
-    ): QueryBuilder {
-        $qb = $this->connection->createQueryBuilder();
-        $qb
-            ->addSelect('p.id_product, pl.name, p.reference, i.id_image')
-            ->from($this->dbPrefix . 'product', 'p')
-            ->join('p', $this->dbPrefix . 'product_shop', 'ps', 'ps.id_product = p.id_product AND ps.id_shop = :shopId')
-            ->leftJoin('p', $this->dbPrefix . 'product_lang', 'pl', 'pl.id_product = p.id_product AND pl.id_lang = :languageId')
-            ->leftJoin('p', $this->dbPrefix . 'image_shop', 'i', 'i.id_product = p.id_product AND i.id_shop = :shopId AND i.cover = 1')
-            ->leftJoin('p', $this->dbPrefix . 'product_supplier', 'psu', 'psu.id_product = p.id_product')
-            ->leftJoin('p', $this->dbPrefix . 'product_attribute', 'pa', 'pa.id_product = p.id_product')
-            ->setParameter('shopId', $shopId->getValue())
-            ->setParameter('languageId', $languageId->getValue())
-            ->addOrderBy('pl.name', 'ASC')
-            ->addGroupBy('p.id_product')
-        ;
-
-        $dbSearchPhrase = sprintf('"%%%s%%"', $searchPhrase);
-        $qb->where($qb->expr()->or(
-            $qb->expr()->like('pl.name', $dbSearchPhrase),
-
-            // Product references
-            $qb->expr()->like('p.isbn', $dbSearchPhrase),
-            $qb->expr()->like('p.upc', $dbSearchPhrase),
-            $qb->expr()->like('p.mpn', $dbSearchPhrase),
-            $qb->expr()->like('p.reference', $dbSearchPhrase),
-            $qb->expr()->like('p.ean13', $dbSearchPhrase),
-            $qb->expr()->like('p.supplier_reference', $dbSearchPhrase),
-
-            // Combination attributes
-            $qb->expr()->like('pa.isbn', $dbSearchPhrase),
-            $qb->expr()->like('pa.upc', $dbSearchPhrase),
-            $qb->expr()->like('pa.mpn', $dbSearchPhrase),
-            $qb->expr()->like('pa.reference', $dbSearchPhrase),
-            $qb->expr()->like('pa.ean13', $dbSearchPhrase),
-            $qb->expr()->like('pa.supplier_reference', $dbSearchPhrase)
-        ));
-
-        if (!empty($filters)) {
-            foreach ($filters as $type => $filter) {
-                switch ($type) {
-                    case 'filteredTypes':
-                        $qb->andWhere('p.product_type not in(:filter)')
-                            ->setParameter('filter', implode(', ', $filter));
-                        break;
-                    default:
-                        break;
-                }
-            }
-        }
-
-        if (!empty($limit)) {
-            $qb->setMaxResults($limit);
-        }
-
-        return $qb;
     }
 
     /**
