+++ src/Adapter/Product/Update/RelatedProductsUpdater.php
@@ -28,7 +28,7 @@ declare(strict_types=1);
 
 namespace PrestaShop\PrestaShop\Adapter\Product\Update;
 
-use PrestaShop\PrestaShop\Adapter\Product\Repository\ProductRepository;
+use PrestaShop\PrestaShop\Adapter\Product\Repository\ProductMultiShopRepository;
 use PrestaShop\PrestaShop\Core\Domain\Product\Exception\CannotUpdateProductException;
 use PrestaShop\PrestaShop\Core\Domain\Product\ValueObject\ProductId;
 use PrestaShop\PrestaShop\Core\Exception\CoreException;
@@ -41,15 +41,15 @@ use Product;
 class RelatedProductsUpdater
 {
     /**
-     * @var ProductRepository
+     * @var ProductMultiShopRepository
      */
     private $productRepository;
 
     /**
-     * @param ProductRepository $productRepository
+     * @param ProductMultiShopRepository $productRepository
      */
     public function __construct(
-        ProductRepository $productRepository
+        ProductMultiShopRepository $productRepository
     ) {
         $this->productRepository = $productRepository;
     }
@@ -60,7 +60,7 @@ class RelatedProductsUpdater
      */
     public function setRelatedProducts(ProductId $productId, array $relatedProductIds): void
     {
-        $product = $this->productRepository->get($productId);
+        $product = $this->productRepository->getProductByDefaultShop($productId);
 
         if (empty($relatedProductIds)) {
             $this->deleteRelatedProducts($product);
