+++ src/PrestaShopBundle/Form/Admin/Sell/Product/Pricing/SpecificPriceType.php
@@ -28,6 +28,7 @@ declare(strict_types=1);
 namespace PrestaShopBundle\Form\Admin\Sell\Product\Pricing;
 
 use PrestaShop\PrestaShop\Adapter\Attribute\Repository\AttributeRepository;
+use PrestaShop\PrestaShop\Adapter\Product\Repository\ProductMultiShopRepository;
 use PrestaShop\PrestaShop\Adapter\Product\Repository\ProductRepository;
 use PrestaShop\PrestaShop\Core\ConstraintValidator\Constraints\DateRange;
 use PrestaShop\PrestaShop\Core\Domain\Language\ValueObject\LanguageId;
@@ -61,7 +62,7 @@ class SpecificPriceType extends TranslatorAwareType
     private $urlGenerator;
 
     /**
-     * @var ProductRepository
+     * @var ProductMultiShopRepository
      */
     private $productRepository;
 
@@ -89,13 +90,13 @@ class SpecificPriceType extends TranslatorAwareType
      * @param TranslatorInterface $translator
      * @param array $locales
      * @param UrlGeneratorInterface $urlGenerator
-     * @param ProductRepository $productRepository
+     * @param ProductMultiShopRepository $productRepository
      */
     public function __construct(
         TranslatorInterface $translator,
         array $locales,
         UrlGeneratorInterface $urlGenerator,
-        ProductRepository $productRepository,
+        ProductMultiShopRepository $productRepository,
         AttributeRepository $attributeRepository,
         EventSubscriberInterface $specificPriceCombinationListener,
         CombinationNameBuilderInterface $combinationNameBuilder,
